requirements.txt
chirptext >= 0.1a6
pyyaml
